first refacto of wrap tests
