+++ crate/server/src/routes/ms_dke/mod.rs
@@ -15,7 +15,7 @@ use cosmian_kmip::kmip::{
 };
 use num_bigint_dig::BigUint;
 use serde::{Deserialize, Serialize};
-use tracing::{info, trace};
+use tracing::info;
 use url::Url;
 
 use crate::{kms_bail, kms_error, result::KResult, KMSServer};
@@ -241,7 +241,6 @@ fn big_uint_to_u32(bu: &BigUint) -> u32 {
 
 #[cfg(test)]
 mod tests {
-    use base64::{engine::general_purpose::STANDARD, Engine};
     use num_bigint_dig::BigUint;
 
     use crate::routes::ms_dke::big_uint_to_u32;
@@ -259,13 +258,5 @@ mod tests {
         let bu = BigUint::from(1_u64 << 32);
         assert_eq!(5, bu.to_bytes_be().len());
         assert_eq!(0, big_uint_to_u32(&bu));
-    }
-
-    #[test]
-    fn test_decrypt_sample() {
-        let b64_string = "wP4ir0aynve6Cpv3ZcBo5+HDue7OA6ogQetNkql1ptfKXilQ2N6x+wDTszcrJlb672l+ckUV5Gjn+ohhFUh0hx6B3rTNKVyxJiGq8S+MRXrTl0UGjWjFED7fYZ2nYZPigu1VHdm3HgBVZdeR8TMr1uIjDHxhWgen2utnTvacn5r8X079ImwpbhilrYBUvt9q42r/CxRp+axsMY3ozkGYsSZ/vXsgjSN0Nbn+9cwHi+XeE2PcjAOnaxUTKVcxjcZvRE+y2FcwgT+nVfJub4ZvRjz9lAbhdDNUS2ZrisAtHVRWJx1ArAMHH7OYg41LoA9+wmBoB04cEzi3JkJkqNCwtw==";
-        let bytes = STANDARD.decode(b64_string.as_bytes()).unwrap();
-        let bu = BigUint::from_bytes_be(&bytes);
-        // println!("bu: {:?}", bu.to_str_radix(10_u32));
     }
 }
