+++ crate/utils/src/crypto/wrap/rsa_oaep_aes_kwp.rs
@@ -4,7 +4,7 @@ use openssl::{
     rand::rand_bytes,
 };
 
-use super::rfc5649::{key_unwrap, key_wrap, AES_KWP_KEY_SIZE};
+use super::rfc5649::{rfc_5649_unwrap, tfc_5649_wrap, AES_KWP_KEY_SIZE};
 use crate::{
     crypto::wrap::ckm_rsa_pkcs_oaep::{
         ckm_rsa_pkcs_oaep_key_unwrap, ckm_rsa_pkcs_oaep_key_wrap, RsaOaepHash,
@@ -46,7 +46,7 @@ pub fn ckm_rsa_aes_key_wrap(
     let encapsulation = ckm_rsa_pkcs_oaep_key_wrap(pubkey, hash_fn, kek.clone())?;
 
     // Wrap key according to RFC 5649 as recommended.
-    let wk = key_wrap(plaintext, &kek)?;
+    let wk = tfc_5649_wrap(plaintext, &kek)?;
 
     Ok([encapsulation, wk].concat())
 }
@@ -101,7 +101,7 @@ pub fn ckm_rsa_aes_key_unwrap(
     let kek = ckm_rsa_pkcs_oaep_key_unwrap(p_key, hash_fn, encapsulation)?;
 
     // Unwrap key according to RFC 5649 as recommended.
-    let plaintext = key_unwrap(wk, &kek)?;
+    let plaintext = rfc_5649_unwrap(wk, &kek)?;
 
     Ok(plaintext)
 }
