+++ crate/utils/src/kmip_utils.rs
@@ -3,7 +3,7 @@ use zeroize::Zeroizing;
 use crate::{
     crypto::{
         password_derivation::derive_key_from_password,
-        wrap::{key_unwrap, key_wrap},
+        wrap::{rfc_5649_unwrap, tfc_5649_wrap},
     },
     error::KmipUtilsError,
 };
@@ -14,7 +14,7 @@ const WRAPPING_SECRET_LENGTH: usize = 32;
 pub fn wrap_key_bytes(key: &[u8], wrapping_password: &str) -> Result<Vec<u8>, KmipUtilsError> {
     let wrapping_secret =
         derive_key_from_password::<WRAPPING_SECRET_LENGTH>(wrapping_password.as_bytes())?;
-    key_wrap(key, wrapping_secret.as_ref()).map_err(|e| KmipUtilsError::Default(e.to_string()))
+    rfc_5649_wrap(key, wrapping_secret.as_ref()).map_err(|e| KmipUtilsError::Default(e.to_string()))
 }
 
 /// Unwrap a key using a password
@@ -24,5 +24,5 @@ pub fn unwrap_key_bytes(
 ) -> Result<Zeroizing<Vec<u8>>, KmipUtilsError> {
     let wrapping_secret =
         derive_key_from_password::<WRAPPING_SECRET_LENGTH>(wrapping_password.as_bytes())?;
-    key_unwrap(key, wrapping_secret.as_ref()).map_err(|e| KmipUtilsError::Default(e.to_string()))
+    rfc_5649_unwrap(key, wrapping_secret.as_ref()).map_err(|e| KmipUtilsError::Default(e.to_string()))
 }
